remove unused inference_transpiler unit-tests
test=develop
